move config struct and methods to own package
